% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_in_html.R
\name{note_in_html}
\alias{note_in_html}
\title{Print text input in html}
\usage{
note_in_html(input = NULL)
}
\arguments{
\item{input}{tags id of textarea from shiny ui}
}
\description{
Print text input in html
}
\examples{
note_in_html("Please write note using **Markdown** syntax! ")
}
