% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_in_md_pdf.R
\name{note_in_md_pdf}
\alias{note_in_md_pdf}
\title{Print text input in pdf}
\usage{
note_in_md_pdf(input = NULL)
}
\arguments{
\item{input}{tags id of textarea from shiny ui}
}
\description{
Print text input in pdf
}
\examples{
note_in_md_pdf("Please write note using **Markdown** syntax! ")
}
