% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_in_md_word.R
\name{note_in_md_word}
\alias{note_in_md_word}
\title{Print text input in word}
\usage{
note_in_md_word(input = NULL)
}
\arguments{
\item{input}{tags id of textarea from shiny ui}
}
\description{
Print text input in word
}
\examples{
note_in_md_word("Please write note using **Markdown** syntax! ")
}
