% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_in_md_pdf.R
\name{note_in_md_pdf}
\alias{note_in_md_pdf}
\title{Title}
\usage{
note_in_md_pdf(input)
}
\arguments{
\item{input}{}
}
\description{
Title
}
\examples{
note_in_md_pdf(input$markdownid)
}
