% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_in_md_word.R
\name{note_in_md_word}
\alias{note_in_md_word}
\title{Title}
\usage{
note_in_md_word(input)
}
\arguments{
\item{input}{}
}
\description{
Title
}
\examples{
note_in_md_word(input$markdownid)
}
